requirements.txt
discord.py>=2.0.0
sqlite3
logging
flask
random
string